// admin-clients/libs/cpanel/venues/feature/standard-venue-tpl-manager/src/lib/models/venue-template-editor-type.model.ts
export enum VenueTemplateEditorType {
    venueTemplate = 'venueTemplate',
    promoterTemplate = 'promoterTemplate',
    eventTemplate = 'eventTemplate',
    sessionTemplate = 'sessionTemplate',
    sessionPackTemplate = 'sessionPackTemplate',
    multiSessionTemplate = 'multiSessionTemplate',
    seasonTicketTemplate = 'seasonTicketTemplate'
}